Integrate load cell.
Get a moving average of the weight values measured by the load cell.
The size of the moving average window is kept small to weed out erratic
changes due to noise or whatever in the load cell's (occassionally)
unreliable readings.

A "timed" moving average module is created which an internal millis()
based simple timing logic to report the updated moving average weight
to the main task.